The sentiment return mapping notebook
This file first uses simple text CNN to give a positive prob and predicted label to the stock file base on the desired date. then it extracts the stock return(after stock data preparation) from the stock excels and maps them with the respected stock and date in sentiment analysis. It generates two excel files, one is sentiment analysis and the other is the respected stock return. After that one can simply use lookup() function in excel to combine the data.